refactor: split settings constants by domain
- Move code constants into domain modules under src/settings
- Keep constants.ts as a re-export barrel for existing import sites
- Update configuration.md header and restore documented constant rows
- Stop exporting actedThreadIdsFromDetail after knip flagged it unused
