+++ src/settings/context7Constants.ts
@@ -1,0 +1,2 @@
+/** Context7 integration. */
+export const CONTEXT7_BASE_URL = "https://context7.com/api";
