+++ src/settings/cursorConstants.ts
@@ -1,0 +1,10 @@
+/** pi-ai metadata when mapping Cursor.models.list() items. */
+export const CURSOR_DEFAULT_CONTEXT_WINDOW = 200_000;
+export const CURSOR_DEFAULT_MAX_TOKENS = 16_384;
+
+/** Cursor SDK MCP bridge (inline provider). */
+export const CURSOR_MCP_BIND_HOST = "127.0.0.1";
+export const CURSOR_MCP_TOKEN_BYTES = 32;
+export const CURSOR_MCP_SERVER_START_TIMEOUT_MS = 5_000;
+export const CURSOR_MAX_PORT_RETRIES = 5;
+export const CURSOR_MCP_SERVER_NAME = "pr-agent";
