+++ src/settings/githubConstants.ts
@@ -1,0 +1,7 @@
+/** GitHub API / token handling. */
+export const PRIMARY_RATE_LIMIT_MAX_RETRIES = 2;
+export const SECONDARY_RATE_LIMIT_MAX_RETRIES = 3;
+
+export const GITHUB_REACTION_EYES = "eyes" as const;
+
+export const DEFAULT_COOLDOWN_SECONDS = 60;
