+++ src/settings/loggingConstants.ts
@@ -1,0 +1,3 @@
+/** Logging. */
+export const MAX_LOG_MESSAGE_LEN = 2_000;
+export const MAX_LOG_REDACTION_SCAN_LEN = MAX_LOG_MESSAGE_LEN * 4;
