+++ src/settings/paginationConstants.ts
@@ -1,0 +1,5 @@
+export const GITHUB_PULL_REQUEST_FILES_API_MAX_FILES = 3_000;
+export const COMMENTS_PAGE_SIZE = 100;
+export const COMMENT_PAGINATION_MAX_PAGES = 20;
+export const PR_COMMITS_PAGE_SIZE = 100;
+export const PR_COMMITS_MAX_PAGES = 20;
