+++ src/settings/slashConstants.ts
@@ -1,0 +1,22 @@
+/** Slash command help (scheduler ack replies). */
+export const SLASH_HELP_BODY = [
+  "### PR Agent help",
+  "",
+  "Commands (first line of a **new** comment):",
+  "- `/help` — show this message",
+  "- `/ask <question>` — ask about this PR or a specific line of code",
+  "- `/describe` — generate or refresh the PR title/body summary (also runs automatically on PR open)",
+  "- `/review` — general bug-and-correctness review (also runs automatically on PR open; further reviews need a manual `/review`)",
+  "- `/review-security` — deep security review (DeepSec-style; trigger-only, not auto-run)",
+  "- `/review-quality` — deep code-quality review (maintainability; trigger-only, not auto-run)",
+  "- `/review-tests` — draft missing test cases for the PR's changes (trigger-only, not auto-run)",
+  "- `/triage` — fix earlier PR Agent findings on this PR: commits and pushes minimal fixes to the PR branch, resolves fixed threads (trigger-only; same-repo PRs). Post on the PR conversation to triage all findings, or reply `/triage` inside a bot inline finding thread to triage that finding only.",
+  "",
+  "Notes:",
+  "- Automated `/describe` runs on PR actions listed in `DESCRIPTION_AUTO_ACTIONS` (default `opened` only); `/review` runs on PR actions listed in `REVIEW_AUTO_ACTIONS` (default `opened` only, so follow-up pushes need a manual `/review`).",
+  "- `/describe` merges generated content below the PR Agent description header; your text above that header is preserved.",
+  "- `/review`, `/review-security`, `/review-quality`, and `/review-tests` can each leave summary comments on the same PR (different sentinels).",
+  "- `/ask` answers one question at a time; it does not remember prior `/ask` commands.",
+  "- Some security issues may appear in both passes; pick the command that matches your question.",
+  "- Edited comments are ignored for slash parsing in v1.",
+].join("\n");
