+++ src/settings/workspaceConstants.ts
@@ -1,0 +1,3 @@
+export const LOCAL_WORKSPACE_TREE_WALK_CONCURRENCY = 32;
+export const LOCAL_WORKSPACE_GREP_PATHSPEC_CHUNK_SIZE = 256;
+export const PR_REPOSITORY_VIEW_RELEASE_GRACE_MS = 60_000;
